SurfaceFlinger: Allow forcing faking of Vsync
Not all displays will have a vsync like HDMI, TV or
manual update displays.

The orginal code assumes that just because the HWComposer HAL
has a certain API version that the HW Vsync should be supported.

But that is not always true.
This patch allows a system property that will allow forcing SW vsync
based on system property sf.force_sw_vsync.  This will allow devices
that do not support Vsync or processor HALs that have this
API version the ability to use the SW Vsync thread.

Change-Id: I45ac7aa81d4d132b04bf8b48823a2657eeffd683
Signed-off-by: Dan Murphy <dmurphy@ti.com>