egl: Fix debug flag for SW composition in Surfaceflinger
The debug.sf.hw property makes it possible to force Surfaceflinger
composition in SW by making it choose a slow EGL config. This feature
was broken in JB, which always loads only the HW renderer.
This change fixes it by loading the SW renderer when Surfaceflinger is
started while the debug property is set to 0.

Change-Id: Ibe44c1e6e7c9059d7be024c36121a42a25450aef