surfaceflinger: fix inverse transform calculation.
there is a bug in SurfaceFlinger when calculating the
inverse transform of FLIP_H|ROT_90 and of FLIP_V|ROT_90.

it incorrectly calculates that inverse transform of FLIP_H|ROT_90
is FLIP_V|ROT_90 when in fact it should be FLIP_H|ROT_90.
(note: in android, flips are applied before rotation).

the inverse transform of FLIP_V|ROT_90 is again incorrectly
calculated as FLIP_H|ROT_90, when in fact it should be
FLIP_V|ROT_90.

b/11914082

Change-Id: I6a6bef0342322dbb422f2a7247584a78921a4d00
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>